% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAerreg.R
\name{DAer}
\alias{DAer}
\title{Data Augmentation for censored expectile regression}
\usage{
DAer(
  dataset,
  imptau = seq(0.01, 0.99, 0.01),
  tau = seq(0.1, 0.9, 0.1),
  H = 20,
  censortype = c("right", " left", "interval"),
  formula
)
}
\arguments{
\item{dataset}{A censoring dataset}

\item{imptau}{The imputation expectile. The default value isn\{0.01,...0.09\}}

\item{tau}{A set of expectile levels of interest}

\item{H}{Iteration time, default H = 20}

\item{censortype}{Censoring  mechanism, 'right' for right censoring, 'left' for left censoring, 'interval' for interval censoring.}

\item{formula}{A formula expression for regression model}
}
\value{
A list object, which is basically a list consisting of:
\item{finalbeta}{Parameter estimation of interest}
\item{finalyhat}{Fitted value}
}
\description{
Data Augmentation for censored expectile regression
}
