% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAerreg.R
\name{DAer_ensemble}
\alias{DAer_ensemble}
\title{Combines the estimators from multiple iterations to produce a robust and aggregated estimate.}
\usage{
DAer_ensemble(betahat, yhat, H)
}
\arguments{
\item{betahat}{estimators}

\item{yhat}{fitted value}

\item{H}{iteration time}
}
\description{
Combines the estimators from multiple iterations to produce a robust and aggregated estimate.
}
