% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAerreg.R
\name{DAer_est}
\alias{DAer_est}
\title{Function for Data Augmentation}
\usage{
DAer_est(
  dataset,
  impY,
  Kn = seq(0.1, 0.9, 0.1),
  imptau = seq(0.01, 0.99, 0.01),
  formula
)
}
\arguments{
\item{dataset}{A censoring dataset}

\item{impY}{Dependent value after imputation}

\item{Kn}{A set of expectile levels of interest}

\item{imptau}{The imputation expectile. The default value isn\{0.01,...0.09\}}

\item{formula}{A formula expression for regression model}
}
\value{
A list object, which is basically a list consisting of:
\item{bhat}{Parameter estimation of interest}
\item{yhat}{Fitted value}
\item{intbeta}{The updata imputation parameters}
}
\description{
Function for Data Augmentation
}
