% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAerreg.R
\name{DAer_resemble}
\alias{DAer_resemble}
\title{Resemble iteration}
\usage{
DAer_resemble(betahat, yhat, H)
}
\arguments{
\item{betahat}{estimate}

\item{yhat}{fitted value}

\item{H}{itertation time}
}
\description{
Resemble iteration
}
