% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagenerate.R
\name{data.generate}
\alias{data.generate}
\title{Title Function for generating simulation data}
\usage{
data.generate(
  n,
  xdis,
  edis,
  sd = 0.5,
  tau = seq(0.1, 0.9, 0.1),
  beta0,
  beta1,
  gamma0,
  gamma1,
  censor.type,
  SNR,
  c0 = 0,
  l0 = -0.3,
  rate
)
}
\description{
Title Function for generating simulation data
}
