% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAerreg.R
\name{impy}
\alias{impy}
\title{Function for imputation censoring samples}
\usage{
impy(dataset, beta, censor.type = c("right", " left", "interval"))
}
\arguments{
\item{dataset}{A censoring dataset}

\item{beta}{The imputation parameters}

\item{censor.type}{Censoring  mechanism, 'right' for right censoring, 'left' for left censoring, 'interval' for interval censoring}
}
\value{
Imputation value for censoring samples
}
\description{
Return the imputaion value for censoring samples
}
