% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAerreg.R
\name{intBeta}
\alias{intBeta}
\title{Function for parameters initialization}
\usage{
intBeta(dataset, imptau = seq(0.01, 0.99, 0.01), formula)
}
\arguments{
\item{dataset}{A censoring dataset}

\item{imptau}{The imputation expectile. The default value isn\{0.01,...0.09\}}

\item{formula}{A formula expression for regression model}
}
\value{
Initialized imputation parameters
}
\description{
Return initialized imputation parameters, the expectile parameter is estimated under common slope assumption.
}
