% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAerreg.R
\name{DAer}
\alias{DAer}
\title{Implements the data augmentation steps specific to expectile regression.}
\usage{
DAer(
  dataset,
  imptau = seq(0.01, 0.99, 0.01),
  tau = seq(0.1, 0.9, 0.1),
  H = 20,
  censortype = c("right", " left", "interval"),
  formula
)
}
\arguments{
\item{dataset}{A censoring dataset}

\item{imptau}{A sequence of expectile levels used for imputing censored values. The default is a dense sequence from 0.01 to 0.99.}

\item{tau}{The expectile level, representing the target point on the response distribution.}

\item{H}{The total number of iterations for the algorithm. The default value is set to 20.}

\item{censortype}{Censoring  mechanism, 'right' for right censoring, 'left' for left censoring, 'interval' for interval censoring.}

\item{formula}{The regression formula, consisting of the response variable, '~' and the sum of all effects that should be taken into consideration.}
}
\value{
A list object, which is basically a list consisting of:
\item{finalbeta}{Parameter estimation at interest $tau$ level}
\item{finalyhat}{Fitted value}
}
\description{
Implements the data augmentation steps specific to expectile regression.
}
