% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAerreg.R
\name{impy}
\alias{impy}
\title{Computes and returns the imputed values for censored samples.}
\usage{
impy(dataset, beta, censor.type = c("right", " left", "interval"))
}
\arguments{
\item{dataset}{A censoring dataset}

\item{beta}{The imputation parameters}

\item{censor.type}{The censoring mechanism. Options include right-censoring, left-censoring, and interval-censoring.}
}
\value{
Imputed value for censoring samples
}
\description{
Return the imputation value for censoring samples
}
